refactor, move util functions out of main.go
