+++ utils.go
@@ -1,6 +1,12 @@
 package main
 
-import "github.com/mitchellh/mapstructure"
+import (
+	"crypto/sha256"
+	"encoding/hex"
+	"strings"
+
+	"github.com/mitchellh/mapstructure"
+)
 
 func mapToObj(input any, output any) error {
 	cfg := &mapstructure.DecoderConfig{
@@ -17,4 +23,29 @@ func mapToObj(input any, output any) error {
 
 	decoder.Decode(input)
 	return nil
+}
+
+func parseHeaders(headers []string) map[string]string {
+	// Parse headers into a map
+
+	headersMap := make(map[string]string)
+
+	for _, header := range headers {
+		splittedHeader := strings.SplitN(strings.ReplaceAll(header, " ", ""), ":", 2)
+		headerKey := splittedHeader[0]
+		headerValue := splittedHeader[1]
+
+		headersMap[headerKey] = headerValue
+	}
+
+	return headersMap
+}
+
+func sha256encode(val []byte) string {
+	h := sha256.New()
+
+	h.Write([]byte(val))
+
+	bs := h.Sum(nil)
+	return hex.EncodeToString(bs)
 }
